rename component to match naming convention
